backend/requirements.txt
fastapi
uvicorn[standard]
mercadopago
python-dotenv
pydantic
starlette
cloudinary
python-multipart
psycopg2-binary